Odradjeni Servisi i Repository

+++ src/main/java/com/grupaF/restoran/models/Porudzbina.java
@@ -10,7 +10,7 @@ import java.util.Set;
 public class Porudzbina {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int iDPorudzbina;
+    private Long iDPorudzbina;
 
     @Column(name = "otkazano")
     private boolean otkazano;
@@ -28,11 +28,11 @@ public class Porudzbina {
     @OneToMany(mappedBy = "porudzbina")
     Set<Porudzbina_proizvod> porudzbina_proizvods;
 
-    public int getiDPorudzbina() {
+    public Long getiDPorudzbina() {
         return iDPorudzbina;
     }
 
-    public void setiDPorudzbina(int iDPorudzbina) {
+    public void setiDPorudzbina(Long iDPorudzbina) {
         this.iDPorudzbina = iDPorudzbina;
     }
